projeto-python-postgresql-analytics/requirements.txt
pandas
matplotlib
sqlalchemy
psycopg[binary]
python-dotenv
